rework popper naming/tracking
the issue with the old counter based method was that different poppers
would get the same name assigned as they finish out of order.

The new logic is actually simpler in that it doesn't discern waiting
from running but just tracks it as busy. It also ensures that workers do
in fact get unique names.