update to current redis client api
their connection pool works really well now, so let's make use of it.
This also gives us a huge increase in stability in case the redis
connection drops.

Now that the connection pool is as reliable, we can get rid of all the
manual connection tracking, so the code gets a lot simpler too